Move most of LoginActivity's onAuthFlowSuccess to AuthenticationUtlities onAuthFlowComplete so code is not repeated for Native Login or IDP.
